# packages/contracts/src/homebrew_mlflow/contracts/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "Homebrew MLflow API",
    "version": "0.1.0"
  },
  "servers": [{"url": "/"}],
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "listProjects", "summary": "List the caller's Research Projects", "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Research Projects", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ProjectResponse"}}}}}}
      },
      "post": {
        "operationId": "createProject",
        "summary": "Create a Research Project and its seeded default repository",
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateProjectRequest"}}}},
        "responses": {
          "202": {"description": "Project and repository provisioning accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectCreationResponse"}}}},
          "403": {"description": "Organization Admin required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
          "409": {"description": "Project slug conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "delete": {
        "operationId": "archiveProject", "summary": "Archive an active Research Project", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Archived project", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectResponse"}}}}, "403": {"description": "Maintainer required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}, "409": {"description": "Invalid lifecycle state", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
    "/api/v1/projects/{project_id}/restore": {
      "post": {
        "operationId": "restoreProject", "summary": "Restore an archived Research Project as an organization Admin", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Restored project", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectResponse"}}}}, "403": {"description": "Organization Admin required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}, "409": {"description": "Invalid lifecycle state", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
    "/api/v1/projects/{project_id}/repositories": {
      "get": {
        "operationId": "listProjectRepositories", "summary": "List hosted project repositories and GitLab links", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Project repositories", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/RepositoryResponse"}}}}}}
      },
      "post": {
        "operationId": "createRepository",
        "summary": "Create and seed another private project repository",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateRepositoryRequest"}}}},
        "responses": {
          "202": {"description": "Repository provisioning accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RepositoryResponse"}}}},
          "403": {"description": "Project Maintainer required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
          "409": {"description": "Repository slug conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}
        }
      }
    },
    "/api/v1/projects/{project_id}/repositories/{repository_id}": {
      "delete": {
        "operationId": "archiveRepository", "summary": "Archive a hosted Git repository record", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "repository_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Archived Repository", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RepositoryResponse"}}}}, "403": {"description": "Maintainer required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "schemas": {
...
      "CreateProjectRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["organization_id", "name", "slug"],
        "properties": {
          "organization_id": {"type": "string", "pattern": "^org_[0-9A-HJKMNP-TV-Z]{26}$"},
          "name": {"type": "string", "minLength": 1, "maxLength": 200},
          "slug": {"type": "string", "minLength": 1, "maxLength": 100}
        }
      },
      "CreateRepositoryRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["name", "slug"],
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 200},
          "slug": {"type": "string", "minLength": 1, "maxLength": 100},
          "default_branch": {"type": "string", "default": "main", "minLength": 1, "maxLength": 100}
        }
      },
...
      "RepositoryResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "project_id", "name", "slug", "default_branch", "state", "web_url", "http_clone_url", "ssh_clone_url", "failure_code"],
        "properties": {
          "id": {"type": "string", "pattern": "^repo_[0-9A-HJKMNP-TV-Z]{26}$"},
          "project_id": {"type": "string", "pattern": "^pr_[0-9A-HJKMNP-TV-Z]{26}$"},
          "name": {"type": "string"}, "slug": {"type": "string"},
          "default_branch": {"type": "string"}, "state": {"type": "string", "enum": ["provisioning", "active", "failed", "archived"]},
          "web_url": {"type": ["string", "null"]}, "http_clone_url": {"type": ["string", "null"]},
          "ssh_clone_url": {"type": ["string", "null"]}, "failure_code": {"type": ["string", "null"]}
        }
      },
...
      "ProjectCreationResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "organization_id", "name", "slug", "default_repository"],
        "properties": {
          "id": {"type": "string", "pattern": "^pr_[0-9A-HJKMNP-TV-Z]{26}$"},
          "organization_id": {"type": "string", "pattern": "^org_[0-9A-HJKMNP-TV-Z]{26}$"},
          "name": {"type": "string"}, "slug": {"type": "string"},
          "default_repository": {"$ref": "#/components/schemas/RepositoryResponse"}
        }
      },
      "ProjectResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "organization_id", "name", "slug", "state", "created_at", "archived_at"],
        "properties": {"id": {"type": "string"}, "organization_id": {"type": "string"}, "name": {"type": "string"}, "slug": {"type": "string"}, "state": {"type": "string"}, "created_at": {"type": "string", "format": "date-time"}, "archived_at": {"type": ["string", "null"], "format": "date-time"}}
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "request_id", "details", "retryable"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "request_id": {"type": "string"},
          "details": {"type": "object"},
          "retryable": {"type": "boolean"}
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {"error": {"$ref": "#/components/schemas/Error"}}
      }
    }
  }
}